move isghash to bt pkg
