+++ bt/SignatureHash_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 
 	"github.com/libsv/libsv/bt"
-	"github.com/libsv/libsv/bt/sig/sighash"
+	"github.com/libsv/libsv/bt/sighash"
 	"github.com/libsv/libsv/script"
 )
