+++ bt/sighash/SigHashFlag.go
